Optimizer: convert the AccessDumper from a pass to a test
