+++ test/SILOptimizer/accessutils_raw_unit.sil
@@ -1,4 +1,4 @@
-// RUN: %target-sil-opt %s -dump-access -o /dev/null | %FileCheck %s
+// RUN: %target-sil-opt %s -test-runner -o /dev/null | %FileCheck %s
 
 // REQUIRES: swift_in_compiler
 
@@ -36,6 +36,7 @@ struct MoveOnlyPair: ~Copyable {
 // CHECK-NEXT: End accesses for testMarkUnresolvedNonCopyableValueInst
 sil [ossa] @testMarkUnresolvedNonCopyableValueInst : $@convention(thin) (@guaranteed List) -> () {
 bb0(%0 : @guaranteed $List):
+  specify_test "access_path"
   %1 = alloc_box ${ var MoveOnlyPair }
   %2 = project_box %1 : ${ var MoveOnlyPair }, 0
   %3 = mark_unresolved_non_copyable_value [consumable_and_assignable] %2 : $*MoveOnlyPair
